fix: moved UIToolkit code to Own namespace, and folder.
Rearranged the folder structure to leave ourselves open to future UI hookup.
