+++ Assets/Scripts/Friends.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: cb6cb687d8e8f2d438e058123fc8683a
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
