+++ Assets/Scripts/UI/UIToolkit/LocalPlayerView.cs
@@ -4,7 +4,7 @@ using Unity.Services.Friends.Models;
 using UnityEngine;
 using UnityEngine.UIElements;
 
-namespace UnityGamingServicesUsesCases.Relationships.UI
+namespace UnityGamingServicesUsesCases.Relationships.UIToolkit
 {
     public class LocalPlayerView
     {
