+++ Assets/Scripts/Utilities/ApplicationQuit.cs
