+++ Assets/Scripts/Utilities/ApplicationQuit.cs.meta
