+++ Assets/Scripts/Utilities/RelationshipUITester.cs
@@ -3,7 +3,7 @@ using Unity.Services.Friends.Models;
 using UnityEngine;
 
 
-namespace UnityGamingServicesUsesCases.Relationships.UI
+namespace UnityGamingServicesUsesCases.Relationships.UIToolkit
 {
     public class RelationshipUITester : MonoBehaviour
     {
