+++ Assets/Scripts/Views/Debug.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 02a798c41cebfec42bba33d159685a47
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
