+++ Assets/Scripts/Views/Debug/LogInView.cs
